Split window.rs into a window/ module along existing seams
window.rs was 4,777 lines. Split it along the seams the 2026-06-09 review
called out, with no behavior change:

- window/state.rs — the pure reducer cluster: WindowState + its helpers,
  the system-refinement consts, Effect + its Debug impl, the apply()
  reducer, filter_messages, and the full ~1,800-line test module. (Tracked
  as the rename of window.rs so the reducer/test history is preserved.)
- window/voice.rs — wire_voice_controls, wire_embedded_mic, speak_text.
- window/mod.rs — layout (AdelieWindow::new), WindowWidgets, the
  handle_ui_message executor, ensure_active_conversation, now_epoch_ms,
  install_app_icon.

Cross-module items are exposed as pub(super) (WindowState + the four fields
/ two helpers the executor and closures read, Effect, apply,
refinement_for_send, voice_mode_client_tools, the voice wiring fns).
Behavior-preserving: identical default + no-default-features builds, clippy
clean, and all 225/224 tests pass.
